dev-requirements.txt
# Testing
pytest
pytest-django
coverage

# API Documentation
drf-yasg

# For development
# pip install -r requirements-dev.txt

# For production
# pip install -r requirements.txt